Separate definitions and implementations into individual code files
Hopefully, this will make the code easier to digest, learn from,
and maintain (moving forward).
