+++ src/Examples/Behavioral/TemplateMethod/TemplateMethodDemo.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Examples.Behavioral.TemplateMethod
+﻿namespace Examples.Behavioral.TemplateMethod
 {
     public class TemplateMethodDemo : DemoBase
     {
