+++ src/Examples/Behavioral/Visitor/Circle.cs
@@ -1,0 +1,10 @@
+﻿namespace Examples.Behavioral.Visitor
+{
+    public class Circle : Shape
+    {
+        public override void Accept(IShapeVisitor visitor)
+        {
+            visitor.Visit(this);
+        }
+    }
+}
