+++ src/Examples/Behavioral/Visitor/IShapeVisitor.cs
@@ -1,0 +1,11 @@
+﻿namespace Examples.Behavioral.Visitor
+{
+    public interface IShapeVisitor
+    {
+        void Visit(Circle circle);
+
+        void Visit(Square square);
+
+        void Visit(Triangle triangle);
+    }
+}
