+++ src/Examples/Behavioral/Visitor/Shape.cs
@@ -1,0 +1,9 @@
+﻿namespace Examples.Behavioral.Visitor
+{
+    public abstract class Shape
+    {
+        public string Name { get; set; }
+
+        public abstract void Accept(IShapeVisitor visitor);
+    }
+}
