+++ src/Examples/Behavioral/Visitor/Square.cs
@@ -1,0 +1,10 @@
+﻿namespace Examples.Behavioral.Visitor
+{
+    public class Square : Shape
+    {
+        public override void Accept(IShapeVisitor visitor)
+        {
+            visitor.Visit(this);
+        }
+    }
+}
