+++ src/Examples/Behavioral/Visitor/Triangle.cs
@@ -1,0 +1,10 @@
+﻿namespace Examples.Behavioral.Visitor
+{
+    public class Triangle : Shape
+    {
+        public override void Accept(IShapeVisitor visitor)
+        {
+            visitor.Visit(this);
+        }
+    }
+}
