+++ src/Examples/Behavioral/Visitor/VisitorDemo.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Examples.Behavioral.Visitor
 {
