+++ src/Examples/Creational/Builder/BuilderDemo.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Examples.Creational.Builder
 {
