+++ src/Examples/Creational/Builder/Circle.cs
@@ -1,0 +1,7 @@
+﻿namespace Examples.Creational.Builder
+{
+    public class Circle : Shape
+    {
+        //public override void Accept(IShapeVisitor visitor)
+    }
+}
