+++ src/Examples/Creational/Builder/RandomizingShapeBuilder.cs
@@ -1,21 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Examples.Creational.Builder
 {
-    public interface IShapeBoardBuilder
-    {
-        void SetName(string name);
-
-        void AddCircle(Circle circle);
-        void AddSquare(Square square);
-        void AddTriangle(Triangle triangle);
-
-        ShapeBoard GetShapeBoard();
-    }
-
     public class RandomizingShapeBuilder : IShapeBoardBuilder
     {
         private readonly ShapeBoard _shapeBoard = new ShapeBoard();
@@ -66,4 +52,3 @@ namespace Examples.Creational.Builder
         }
     }
 }
-
