+++ src/Examples/Creational/Builder/Shape.cs
@@ -1,0 +1,9 @@
+﻿namespace Examples.Creational.Builder
+{
+    public abstract class Shape
+    {
+        public string Name { get; set; }
+
+        //public abstract void Accept(IShapeVisitor visitor);
+    }
+}
