+++ src/Examples/Creational/Builder/ShapeBoard.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace Examples.Creational.Builder
 {
