--- src/Examples/Creational/Builder/Shapes.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Examples.Creational.Builder
-{
-    public abstract class Shape
-    {
-        public string Name { get; set; }
-
-        //public abstract void Accept(IShapeVisitor visitor);
-    }
-
-    public class Circle : Shape
-    {
-        //public override void Accept(IShapeVisitor visitor)
-    }
-
-    public class Square : Shape
-    {
-        //public override void Accept(IShapeVisitor visitor)
-    }
-
-    public class Triangle : Shape
-    {
-        //public override void Accept(IShapeVisitor visitor)
-    }
-}
