+++ src/Examples/Creational/Builder/Square.cs
@@ -1,0 +1,7 @@
+﻿namespace Examples.Creational.Builder
+{
+    public class Square : Shape
+    {
+        //public override void Accept(IShapeVisitor visitor)
+    }
+}
