+++ src/Examples/Creational/Builder/Triangle.cs
@@ -1,0 +1,7 @@
+﻿namespace Examples.Creational.Builder
+{
+    public class Triangle : Shape
+    {
+        //public override void Accept(IShapeVisitor visitor)
+    }
+}
