+++ src/Examples/Structural/Bridge/BridgeDemo.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Examples.Structural.Bridge
+﻿namespace Examples.Structural.Bridge
 {
-    public  class BridgeDemo : DemoBase
+    public class BridgeDemo : DemoBase
     {
         protected override void DoRun()
         {
