+++ src/Examples/Structural/Bridge/IPlayList.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Examples.Structural.Bridge
+{
+    public interface IPlayList
+    {
+        string Title { get; }
+
+        IEnumerable<Song> GetSongs();
+    }
+}
