+++ src/Examples/Structural/Bridge/Song.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Examples.Structural.Bridge
+﻿namespace Examples.Structural.Bridge
 {
     public class Song
     {
